Feature complete

+++ src/app/products/product-shell/product-shell.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { Product } from '../product';
@@ -9,10 +9,13 @@ import { State, getShowProductCode, getCurrentProduct, getProducts, getError } f
 
 import { ProductPageActions } from '../state/actions';
 
+import { ConfirmationService } from 'primeng/api';
+
 @Component({
   selector: 'app-product-shell',
   templateUrl: './product-shell.component.html',
-  styleUrls: ['./product-shell.component.scss']
+  styleUrls: ['./product-shell.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ProductShellComponent implements OnInit {
   displayCode$: Observable<boolean>;
@@ -20,7 +23,7 @@ export class ProductShellComponent implements OnInit {
   products$: Observable<Product[]>;
   errorMessage$: Observable<string>;
 
-  constructor(private store: Store<State>) { }
+  constructor(private store: Store<State>, private confirmationService: ConfirmationService) { }
 
   ngOnInit(): void {
 
@@ -53,7 +56,14 @@ export class ProductShellComponent implements OnInit {
   }
 
   deleteProduct(product: Product): void {
-    this.store.dispatch(ProductPageActions.deleteProduct({ productId: product.id }));
+    this.confirmationService.confirm({
+      message: `Are you sure you want to delete the selected (${product.productName}) product?`,
+      header: 'Confirm',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.store.dispatch(ProductPageActions.deleteProduct({ productId: product.id }));
+      }
+    });
   }
 
   clearProduct(): void {
